\name{EnvironmentalA}
\docType{data}
\alias{EnvironmentalA}
\title{Example file for single level environmental factors}
\description{
Example data for Crohns disease in correct input format. Also as text file "EnvironmentalA.txt" in the data folder for this package. data from Calkins, BM (1989).
}
\usage{data("REGENT")}
\format{data frame}
\references{
from Calkins, BM (1989) A meta-analysis of the role of smoking in inflammatory bowel disease
}
\keyword{datasets}
