\name{EnvironmentalB}
\docType{data}
\alias{EnvironmentalB}
\title{Example file for single multiple level environmental factors}
\description{
Example data for Crohns disease in correct input format. Also as text file "EnvironmentalB.txt" in the data folder for this package. *PLEASE NOTE that the multilevel smoking data in EnvironmentalB is entirely artificial.*
}
\usage{data("REGENT")}
\format{data frame}
\keyword{datasets}
