\name{GeneticA}
\docType{data}
\alias{GeneticA}
\title{Example file for SNPs conferring multiplicative risks}
\description{
Example data for Crohns disease in correct input format. Also as text file "GeneticA.txt" in the data folder for this package. Data is from Franke et al (2010), 
}
\usage{data("REGENT")}
\format{data frame}
\references{
Franke et al. (2010), Genome-wide meta-analysis increases to 71 the number of confirmed Crohn's disease susceptibility loci  
}
\keyword{datasets}
