\name{Inds}
\docType{data}
\alias{Inds}
\title{Example of  individual file for REGENT.predict }
\description{
Columns for risk factors, rows for individuals. Entries refer to genotypes (number of risk alleles as defined in the locus file, ie can also be protective alleles) or exposure levels
}
\usage{data("REGENT")}
\format{data frame}
\keyword{datasets}
