\name{REGENT-package}\alias{REGENT}\alias{REGENT-package}\docType{package}\title{Risk Estimation for Genetic and Environmental Traits}\description{Provides risk estimation and categorisation for populations and individuals}\details{\tabular{ll}{Package: \tab REGENT\crType: \tab Package\crVersion: \tab 1.0.6\crDate: \tab 2015-08-18\crLicense: \tab GPL\crLazyLoad: \tab yes\cr}}\author{Daniel Crouch, Graham Goddard & Cathryn Lewis.Maintainer: Daniel Crouch - djmcrouch@gmail.com}\references{Crouch, Goddard & Lewis (2011) Goddard & Lewis, Risk categorization for complex disorders according to genotype relative risk and precision in parameter estimates (2010)}\seealso{\code{\link[REGENT:REGENT.model]{REGENT.model}},\code{\link[REGENT:REGENT.predict]{REGENT.predict}},\code{\link[REGENT:GeneticA]{GeneticA}},\code{\link[REGENT:GeneticB]{GeneticB}},\code{\link[REGENT:EnvironmentalA]{EnvironmentalA}},\code{\link[REGENT:EnvironmentalB]{EnvironmentalB}},\code{\link[REGENT:Inds]{Inds}}}